+++ Document/UserManager.php
@@ -13,91 +13,22 @@ namespace FOS\UserBundle\Document;
 
 use Doctrine\ODM\MongoDB\DocumentManager;
 use FOS\UserBundle\Model\UserInterface;
-use FOS\UserBundle\Model\UserManager as BaseUserManager;
+use FOS\UserBundle\Doctrine\UserManager as BaseUserManager;
 use FOS\UserBundle\Util\CanonicalizerInterface;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
 
+/**
+ * BC class for people extending it in their bundle.
+ * TODO Remove this class on July 31st
+ */
 class UserManager extends BaseUserManager
 {
     protected $dm;
-    protected $repository;
-    protected $class;
 
-    /**
-     * Constructor.
-     *
-     * @param EncoderFactoryInterface $encoderFactory
-     * @param CanonicalizerInterface  $usernameCanonicalizer
-     * @param CanonicalizerInterface  $emailCanonicalizer
-     * @param DocumentManager         $dm
-     * @param string                  $class
-     */
     public function __construct(EncoderFactoryInterface $encoderFactory, CanonicalizerInterface $usernameCanonicalizer, CanonicalizerInterface $emailCanonicalizer, DocumentManager $dm, $class)
     {
-        parent::__construct($encoderFactory, $usernameCanonicalizer, $emailCanonicalizer);
+        parent::__construct($encoderFactory, $usernameCanonicalizer, $emailCanonicalizer, $dm, $class);
 
         $this->dm = $dm;
-        $this->repository = $dm->getRepository($class);
-
-        $metadata = $dm->getClassMetadata($class);
-        $this->class = $metadata->name;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function deleteUser(UserInterface $user)
-    {
-        $this->dm->remove($user);
-        $this->dm->flush();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getClass()
-    {
-        return $this->class;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findUserBy(array $criteria)
-    {
-        return $this->repository->findOneBy($criteria);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function findUsers()
-    {
-        return $this->repository->findAll();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function reloadUser(UserInterface $user)
-    {
-        $this->dm->refresh($user);
-    }
-
-    /**
-     * Updates a user.
-     *
-     * @param UserInterface $user
-     * @param Boolean       $andFlush Whether to flush the changes (default true)
-     */
-    public function updateUser(UserInterface $user, $andFlush = true)
-    {
-        $this->updateCanonicalFields($user);
-        $this->updatePassword($user);
-
-        $this->dm->persist($user);
-        if ($andFlush) {
-            $this->dm->flush();
-        }
     }
 }